requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
google-genai>=1.0
pymupdf>=1.24
pdfplumber>=0.11
requests>=2.31
